Implement 'projects rename' implied project endpoint; refactored shared logic
